+++ src/CleanRegex/Internal/Prepared/Parser/MaskParser.php
@@ -1,30 +1,30 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Prepared\Parser;
 
-use TRegx\CleanRegex\Internal\Prepared\Format\FormatTokenValue;
+use TRegx\CleanRegex\Internal\Prepared\Format\MaskTokenValue;
 use TRegx\CleanRegex\Internal\Prepared\Quotable\Factory\QuotableFactory;
 use TRegx\CleanRegex\Internal\Prepared\Quotable\Quotable;
 
-class FormatParser implements Parser
+class MaskParser implements Parser
 {
     /** @var string */
-    private $format;
+    private $mask;
     /** @var array */
-    private $tokens;
+    private $keywords;
 
-    public function __construct(string $format, array $tokens)
+    public function __construct(string $mask, array $keywords)
     {
-        $this->format = $format;
-        $this->tokens = $tokens;
+        $this->mask = $mask;
+        $this->keywords = $keywords;
     }
 
     public function parse(string $delimiter, QuotableFactory $quotableFactory): Quotable
     {
-        return (new FormatTokenValue($this->format, $this->tokens))->formatAsQuotable();
+        return (new MaskTokenValue($this->mask, $this->keywords))->formatAsQuotable();
     }
 
     public function getDelimiterable(): string
     {
-        return \implode($this->tokens);
+        return \implode($this->keywords);
     }
 }