+++ src/CleanRegex/PatternBuilder.php
@@ -4,8 +4,8 @@ namespace TRegx\CleanRegex;
 use TRegx\CleanRegex\Internal\CompositePatternMapper;
 use TRegx\CleanRegex\Internal\Prepared\Format\IgnoreStrategy;
 use TRegx\CleanRegex\Internal\Prepared\Parser\BindingParser;
-use TRegx\CleanRegex\Internal\Prepared\Parser\FormatParser;
 use TRegx\CleanRegex\Internal\Prepared\Parser\InjectParser;
+use TRegx\CleanRegex\Internal\Prepared\Parser\MaskParser;
 use TRegx\CleanRegex\Internal\Prepared\Parser\PreparedParser;
 use TRegx\CleanRegex\Internal\Prepared\Prepare;
 
@@ -70,9 +70,9 @@ class PatternBuilder
         return new CompositePattern((new CompositePatternMapper($patterns))->createPatterns());
     }
 
-    public function format(string $pattern, array $tokens, string $flags = null): PatternInterface
+    public function mask(string $pattern, array $tokens, string $flags = null): PatternInterface
     {
-        return Prepare::build(new FormatParser($pattern, $tokens), $this->pcre, $flags ?? '');
+        return Prepare::build(new MaskParser($pattern, $tokens), $this->pcre, $flags ?? '');
     }
 
     public function template(string $pattern, string $flags = null): TemplatePattern